Rename allocator namespace to Memory and move files to parent directory
